cico: install kernel-module-internal so we can load mac80211_hwsim

+++ prepare/envsetup.sh
@@ -310,7 +310,12 @@ install_el8_packages () {
     # Install kernel-modules-internal for mac80211_hwsim
     VER=$(rpm -q --queryformat '%{VERSION}' kernel)
     REL=$(rpm -q --queryformat '%{RELEASE}' kernel)
-    dnf -4 -y install http://download.eng.bos.redhat.com/brewroot/vol/rhel-8/packages/kernel/$VER/$REL/$(arch)/kernel-modules-internal-$VER-$REL.$(arch).rpm
+    if ! grep -q -e 'CentOS .* release 8' /etc/redhat-release; then
+        dnf -4 -y install http://download.eng.bos.redhat.com/brewroot/vol/rhel-8/packages/kernel/$VER/$REL/$(arch)/kernel-modules-internal-$VER-$REL.$(arch).rpm
+    else
+        wget https://koji.mbox.centos.org/pkgs/packages/kernel/$VER/$REL/$(arch)/kernel-modules-internal-$VER-$REL.$(arch).rpm --no-check-certificate
+        dnf -y localinstall kernel-modules-internal
+    fi
 
     # Add OVS repo and install OVS
     if ! grep -q -e 'CentOS .* release 8' /etc/redhat-release; then
